+++ openproject-export/lib/open_project/export/engine.rb
@@ -10,15 +10,18 @@ module OpenProject
 
       register 'openproject-export',
                author_url: 'https://openproject.org',
-               requires_openproject: '>= 13.1.0'
-
-      initializer 'openproject-export.register_hooks' do
-        begin
-          ::OpenProject::Export::Hooks
-        rescue NameError
-          # Hooks not loaded (missing OpenProject core)
+               requires_openproject: '>= 13.1.0' do
+        project_module :export_backups do
+          permission :download_project_backup,
+                     { 'open_project/export/backups' => [:download] },
+                     permissible_on: [:project]
         end
       end
+
+      config.to_prepare do
+        require_dependency 'open_project/export/hooks'
+      end
+
     end
   end
 end